==> 3.df_final_experiment_clients.ipynb <==
import pandas as pd

# Specify the file path
file_path = r"C:\Users\cpall\Desktop\ironhack\lessons\5_6_eda_inf_stats_tableau\project\files_for_project\df_final_experiment_clients.txt"

# Read the txt file into a DataFrame
df_roster = pd.read_csv(file_path, sep=",")  # comma-separated values
# Display basic information about the dataframe
df_roster.info()

# # Total count  of clients
total_clients = df_roster["client_id"].nunique()
total_clients

df_roster.columns

# Display the first few rows of the dataframe
df_roster.head(10)

# client_id - Check for duplicates in client_id column: there are none
df_roster.duplicated(subset=['client_id']).sum()

# Total count of clients with Variation data 
total_clients = df_roster["Variation"].nunique()
total_clients

# Frequency table
frequency_table = df_roster['Variation'].value_counts()

# Calculating the proportion of each unique value in the 'process_step'
proportion_table = df_roster['Variation'].value_counts(normalize=True)

frequency_table, proportion_table

x = 26968 +23532
x
